feat(send): fix #19 reduce min message length
We would like to enable users to send messages
containing one or more characters.
The primary motivation for this change is to
support interactions with automated
messaging systems, which often rely on
single-character inputs for responses.

For example:

```
Reply with Y for Yes, N for No, or H for Help.
```